Add CombinedChecker class
This new object class is necessary to handle logging and testing
diferently.
